bot/requirements.txt
requests
pytelegrambotapi
python-decouple
black